% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_color_palette.R
\name{build_color_palette}
\alias{build_color_palette}
\title{Build Color Palette}
\usage{
build_color_palette(df, indexCol)
}
\arguments{
\item{df}{Dataframe of rows to be colored.}

\item{indexCol}{Unquoted name of column to be used for deciding each row's
color.}
}
\value{
\code{df} with an added \code{color} column.
}
\description{
Takes a dataframe and a column to build a color palette for and returns the
dataframe with a \code{color} column added with a unique color for each
category.
}
\examples{
df \%>\% build_color_palette(Category)
}
