% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRrCi.R
\name{calcRrCi}
\alias{calcRrCi}
\title{Calculate a risk ratio confidence interval given a pattern and snp count}
\usage{
calcRrCi(pattern_n, pattern_snp, other_n, other_snp, CI_size = 0.95)
}
\arguments{
\item{pattern_n}{Number of cases with pattern}

\item{pattern_snp}{Number of cases with pattern and SNP minor allele}

\item{other_n}{Number of cases without pattern}

\item{other_snp}{Number of cases without pattern and with SNP minor allele.}

\item{CI_size}{Size of CI in proportion (default 0.95 = 95% interval)}
}
\value{
A list with three entries: PE or RR point estimate, lower and upper for RR lower and upper bounds given interval size.
}
\description{
Calculate a risk ratio confidence interval given a pattern and snp count
}
\examples{
calcRrCi(
 pattern_n=120, pattern_snp=5,
 other_n=2000, other_snp=56,
 CI_size = 0.95
)
}
