% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_data_validation_functions.R
\name{check_genome_file}
\alias{check_genome_file}
\title{Checks genome input dataframe for correct form}
\usage{
check_genome_file(genome, separate = TRUE)
}
\arguments{
\item{genome}{Tibble loaded from supplied file.}

\item{separate}{Return a list with snp name and data? Defaults to \code{TRUE}.}
}
\value{
A list of the genome data with the snp column renamed \code{snp} and
the snp name saved as \code{snp_name}.
}
\description{
This function will lowercase all columns.
}
\examples{
check_genome_file(uploadedGenomeData)
}
