% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_data_validation_functions.R
\name{check_phenome_file}
\alias{check_phenome_file}
\title{Check uploaded phenome file for correctness}
\usage{
check_phenome_file(phenome)
}
\arguments{
\item{phenome}{Uploaded dataframe containing phenotype info}
}
\value{
Dataframe if no errors are thrown.
}
\description{
This function will lowercase all columns.
}
\examples{
check_phenome_file(uploadedPhenomeFile)
}
