% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkgenomefile.R
\name{checkGenomeFile}
\alias{checkGenomeFile}
\title{Checks genome input dataframe for correct form}
\usage{
checkGenomeFile(genome)
}
\arguments{
\item{genome}{Tibble loaded from supplied file.}
}
\value{
A list of the genome data with the snp column renamed \code{snp} and the snp name saved as \code{snp_name}.
}
\description{
Checks genome input dataframe for correct form
}
\examples{
checkGenomeFile(uploadedGenomeData)
}
