% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkphenomefile.R
\name{checkPhenomeFile}
\alias{checkPhenomeFile}
\title{Check uploaded phenome file for correctness}
\usage{
checkPhenomeFile(phenome)
}
\arguments{
\item{phenome}{Uploaded dataframe containing phenotype info}
}
\value{
Dataframe if no errors are thrown.
}
\description{
Check uploaded phenome file for correctness
}
\examples{
checkPhenomeFile(uploadedPhenomeFile)
}
