% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_data_validation_functions.R
\name{checkPhewasFile}
\alias{checkPhewasFile}
\title{Check uploaded phewas file for correctness}
\usage{
checkPhewasFile(phewas)
}
\arguments{
\item{phewas}{Dataframe from uploaded phewas csv}
}
\value{
Supplied dataframe with a tooltip column added if it wasn't
originally present.
}
\description{
Check uploaded phewas file for correctness
}
\examples{
checkPhewasFile('uploadedPhewasFile')
}
