% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseselectedcodes.R
\name{chooseSelectedCodes}
\alias{chooseSelectedCodes}
\title{Choose selected phecodes for app
Takes a selection dataframe (or NULL) and our phewas table and returns the selected phewas codes from the table.}
\usage{
chooseSelectedCodes(selection, phewas_table, just_snps,
  p_value_threshold = 0.001, max_num_phenos_all = 15,
  max_num_phenos_just_snps = 150)
}
\arguments{
\item{selection}{Selection dataframe (NULL will default to a given number of defaults)}

\item{phewas_table}{Phewas table dataframe}

\item{just_snps}{Boolean for if we're in a just snps mode which allows more codes to be seen at a time.}

\item{p_value_threshold}{Float between 0 and 1 for the threshold at which we exclude codes if our selection is too large.}

\item{max_num_phenos_all}{Integer of how max number of phenotypes function will select if including all cases (default = 15).}

\item{max_num_phenos_just_snps}{Integer of max number of phenotypes selected if app is only showing snp MA carriers (default = 150).}
}
\value{
filtered phewas table with either the selection exactly or the most significant snps based upon the \code{max_num_*} limits.
}
\description{
Choose selected phecodes for app
Takes a selection dataframe (or NULL) and our phewas table and returns the selected phewas codes from the table.
}
\examples{
chooseSelectedCodes(NULL, myPhewas, just_snps = FALSE)
}
