% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codefilter.R
\name{codeFilter}
\alias{codeFilter}
\title{Filter codes.}
\usage{
codeFilter(type, code_list, current_codes)
}
\arguments{
\item{type}{Character of 'delete' or 'isolate' deciding which type of filtering is desired}

\item{code_list}{Character vector containing code ids that need to be filtered}

\item{current_codes}{Dataframe with currently selected codes that will be filtered.}
}
\value{
Filtered currently selected codes dataframe.
}
\description{
Filter codes.
}
\examples{
codeFilter('invert', c('001.00', '002.00'), myCurrentCodes)
}
