% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbiditynetwork2d_module.R
\name{comorbidityNetwork2d}
\alias{comorbidityNetwork2d}
\title{Server function of snp info panel}
\usage{
comorbidityNetwork2d(input, output, session, subset_data, results_data,
  inverted_codes, snp_filter, parent_ns)
}
\arguments{
\item{input}{Auto-filled by callModule | ignore}

\item{output}{Auto-filled by callModule | ignore}

\item{session}{Auto-filled by callModule | ignore}
}
\description{
Server function of snp info panel
}
\examples{
callModule(info_panel, 'info_panel', snp_name, individual_data, subset_maf)
}
