% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbiditynetwork2d_module.R
\name{comorbidityNetwork2d_UI}
\alias{comorbidityNetwork2d_UI}
\title{SNP information panel module}
\usage{
comorbidityNetwork2d_UI(id, app_ns)
}
\arguments{
\item{id}{Unique id of module}

\item{app_ns}{A namespace function if this module is included within another module.}
}
\description{
SNP information panel module
}
\examples{

comorbidityNetwork2d_UI('mycomorbidityNetwork2d', parentNS)
}
