% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_loader.R
\name{data_loader_UI}
\alias{data_loader_UI}
\title{Data loading screen: UI}
\usage{
data_loader_UI(id, app_title = "Multimorbidity Explorer")
}
\arguments{
\item{id}{String with unique id of module in app}

\item{app_title}{Name of your app. Defaults to "Multimorbidity Explorer"}
}
\value{
HTML component of shiny module
}
\description{
Data loading screen: UI
}
\examples{
data_loader_UI('my_app')
}
