% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletecodes.R
\name{deleteCodes}
\alias{deleteCodes}
\title{Delete Phecodes from a selected phecodes list with warning.}
\usage{
deleteCodes(codes_to_delete, current_codes)
}
\arguments{
\item{codes_to_delete}{Character vector of codes to delete}

\item{current_codes}{Dataframe containing currently viewed codes with column
\code{code} corresponding to the phecode id.}
}
\value{
Filtered currently viewing codes dataframe.
}
\description{
Delete Phecodes from a selected phecodes list with warning.
}
\examples{
deleteCodes(c('001.00', '002.00'), myCurrentCodes)
}
