% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_id_column.R
\name{detect_id_column}
\alias{detect_id_column}
\title{Detect Id column in data}
\usage{
detect_id_column(data, possible_id_cols = c("iid", "id", "grid"),
  final_id_name = "id")
}
\arguments{
\item{data}{Dataframe that should have id column}

\item{possible_id_cols}{List of possible columns that are acceptable ids.}

\item{final_id_name}{Name of final column that is returned if no errors are
detected.}
}
\value{
Dataframe with a column with name \code{final_id_name} contained,
}
\description{
Takes input data and makes sure that an id column has been provided in proper
format. Used to validate data using \code{\link{checkGenome}}, and
\code{\link{checkPhenome}}.
}
\examples{
dplyr::tibble(
  grid = c('a', 'b', 'c'),
  snp = c(1, 2, 3)
) \%>\% detect_id_column()

}
\seealso{
\code{\link{checkGenome}}, \code{\link{checkPhenome}}
}
