% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_url_state_managers.R
\name{embed_snp_codes_in_url}
\alias{embed_snp_codes_in_url}
\title{Embed snp id and selected codes in app URL}
\usage{
embed_snp_codes_in_url(snp, codes, ma_filtered = FALSE)
}
\arguments{
\item{snp}{Name of current snp}

\item{codes}{List of currently selected codes}

\item{ma_filtered}{Is the subject data currently filtered to minor allele carriers?}
}
\value{
Nothing (updates app URL)
}
\description{
Embed snp id and selected codes in app URL
}
\examples{
embed_snp_codes_in_url('rs123456', c('009.00', '008.10', '008.20'))
}
