% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_url_state_managers.R
\name{extract_snp_codes_from_url}
\alias{extract_snp_codes_from_url}
\title{Extract SNP and Codes from URL query string}
\usage{
extract_snp_codes_from_url(session)
}
\arguments{
\item{session}{session variable from current shiny server. Used to grab current URL string}
}
\value{
A list containing the requested SNP id and codes
}
\description{
Reads a bookmarked state from an apps URL, parses it, and
returns the requested snp and phecodes. Uses \code{\link{extract_snp_codes_from_url_string}} to
parse the string.
}
\examples{
\dontrun{
extract_snp_codes_from_url(session)
}
}
