% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_url_state_managers.R
\name{extract_snp_codes_from_url_string}
\alias{extract_snp_codes_from_url_string}
\title{Extract SNP and Codes from raw URL query string}
\usage{
extract_snp_codes_from_url_string(query_string)
}
\arguments{
\item{session}{session variable from current shiny server. Used to grab current URL string}
}
\value{
A list containing the requested SNP id and codes
}
\description{
Takes a bookmarked state in the form of a url query string, parses it, and
returns the requested snp and phecodes. Called by \code{\link{extract_snp_codes_from_url}}
}
\examples{
extract_snp_codes_from_url_string("?rs123456__00800_90800_08300")
}
