% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_snp_annotations.R
\name{gather_snp_annotations}
\alias{gather_snp_annotations}
\title{Gather snp annotation information}
\usage{
gather_snp_annotations(snp_id)
}
\arguments{
\item{snp_id}{Rsid or name for the SNP}
}
\value{
A two column tibble containing a series of annotation-value pairs for
snp.
}
\description{
Passed a snp rsid or name from the Illumina exome chip, this function returns
a list of pertinant annotations for the user.
}
\examples{
gather_snp_annotations('rs34051416')
}
