% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsnpinfo.R
\name{getSNPInfo}
\alias{getSNPInfo}
\title{Grab snp info from NCBI database
Grabs basic info on a snp from the NCBI for display in the info panel}
\usage{
getSNPInfo(snp)
}
\arguments{
\item{snp}{rsid string of the snp of interest.}
}
\value{
list with values of chromosome, gene, major allele, minor allele, MAF, and base-pair location.
}
\description{
Grab snp info from NCBI database
Grabs basic info on a snp from the NCBI for display in the info panel
}
\examples{
getSNPInf('rs13283456')
}
