% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_help_modal.R
\name{help_modal}
\alias{help_modal}
\title{Help Modal Server}
\usage{
help_modal(input, output, session)
}
\arguments{
\item{input, output, session}{Auto-filled by \code{callModule} | ignore}
}
\description{
Help Modal Server
}
