% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_help_modal.R
\name{help_modal_UI}
\alias{help_modal_UI}
\title{Help Modal UI}
\usage{
help_modal_UI(id, title, help_img_url, more_link)
}
\arguments{
\item{id}{Unique id of module}

\item{title}{Header text for help modal}

\item{help_img_url}{Location of the helper image for modal}

\item{more_link}{Link to further information about context}
}
\description{
Help Modal UI
}
