% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infopanel_module.R
\name{infoPanel}
\alias{infoPanel}
\title{Server function of snp info panel}
\usage{
infoPanel(input, output, session, snp_name, individual_data, subset_maf)
}
\arguments{
\item{input}{Auto-filled by callModule | ignore}

\item{output}{Auto-filled by callModule | ignore}

\item{session}{Auto-filled by callModule | ignore}

\item{snp_name}{String of the snp name}

\item{individual_data}{Individual level snp data dataframe with column \code{snp} containing copies of MA.}

\item{subset_maf}{Minor allele frequency of the current subset of cases.}
}
\description{
Server function of snp info panel
}
\examples{
callModule(info_panel, 'info_panel', snp_name, individual_data, subset_maf)
}
