% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infopanel.R
\name{infoPanel_UI}
\alias{infoPanel_UI}
\title{SNP information panel module}
\usage{
infoPanel_UI(id, app_ns)
}
\arguments{
\item{id}{Unique id of module}

\item{app_ns}{A namespace function if this module is included within another module.}
}
\description{
SNP information panel module
}
\examples{

infoPanel_UI('myInfoPanel', parentNS)
}
