% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertcodes.R
\name{invertCodes}
\alias{invertCodes}
\title{Invert currently selected codes
Only inverts codes that were not preivously inverted. If a code is 'inverted' after it's already
inverted it simple reverts back to normal.}
\usage{
invertCodes(codes_to_invert, currently_inverted_codes)
}
\arguments{
\item{codes_to_invert}{Character vector of codes you wish to invert}

\item{currently_inverted_codes}{Currently selected dataframe with column \code{code} upon which to invert.}
}
\value{
Updated currently selected codes dataframe with previously requested codes inverted from their previous state.
}
\description{
Invert currently selected codes
Only inverts codes that were not preivously inverted. If a code is 'inverted' after it's already
inverted it simple reverts back to normal.
}
\examples{
invertCodes(c('001.00', '002.00'), myCurrentCodes)
}
