% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolatecodes.R
\name{isolateCodes}
\alias{isolateCodes}
\title{Isolate currently viewed codes}
\usage{
isolateCodes(codes_to_isolate, current_codes)
}
\arguments{
\item{codes_to_isolate}{Character vector of codes ids to isolate}

\item{current_codes}{Dataframe with column \code{code} which will be filtered to just rows for requested codes.}
}
\value{
Filtered current codes dataframe.
}
\description{
Isolate currently viewed codes
}
\examples{
isolateCodes(c('001.00', '002.00'), myCurrentCodes)
}
