% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_dashboard.R
\name{main_dashboard}
\alias{main_dashboard}
\title{Server function of upset module}
\usage{
main_dashboard(input, output, session, snp_name, results_data,
  individual_data, MAX_ALLOWED_CODES, COLORS)
}
\arguments{
\item{input, output, session}{Auto-filled by callModule | ignore}
}
\value{
Shiny module
}
\description{
Server function of upset module
}
\examples{
main_dashboard(upset, 'my_mod')
}
