% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_dashboard.R
\name{main_dashboard_UI}
\alias{main_dashboard_UI}
\title{UI function of upset module}
\usage{
main_dashboard_UI(id, div_class = "upset_plot")
}
\arguments{
\item{id}{String with unique id of module in app}
}
\value{
HTML component of shiny module
}
\description{
UI function of upset module
}
\examples{
upset_UI('my_mod')
}
