% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makedescriptionpalette.R
\name{makeDescriptionPalette}
\alias{makeDescriptionPalette}
\title{Create Description Color Palette}
\usage{
makeDescriptionPalette(phecode_info, color_palette = c("#895de6",
  "#6ecc3b", "#e34bca", "#9ad843", "#ff3d97", "#01d8a6", "#ff4254",
  "#64c5ff", "#957100", "#0074d3", "#aad366", "#ff8dcc", "#4e5b00",
  "#ffb38f", "#325c2c", "#faba5f", "#8b373b", "#7e431e"))
}
\arguments{
\item{phecode_info}{Dataframe of phecodes included in the current view of the
application. Must include the column \code{category}.}

\item{color_palette}{Character vector of hex codes to be mapped to the
descriptions in first come first filled order.}
}
\value{
A list containing a palette dataframe and a named array both
containing mappings from description character string to hex code.
}
\description{
Make a standardized color pallete for the phenotype categories
}
\examples{
makeDescriptionPalette(myCurrentPhecodes)
}
