% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makenetworkdata.R
\name{makeNetworkData}
\alias{makeNetworkData}
\title{Convert phenotype dataframe to a data list that works with network plotting functions}
\usage{
makeNetworkData(data, phecode_info, inverted_codes, color_palette,
  case_size = 0.1, code_size = 0.3, no_copies = "#bdbdbd",
  one_copy = "orangered", two_copies = "red")
}
\arguments{
\item{data}{A phenome dataframe with phecodes as columns and also with an IID column and a snp status column}

\item{phecode_info}{Information about each codes results in phewas study in form of columns: code, category, tooltip.}

\item{inverted_codes}{An array of codes that are inverted.}

\item{color_palette}{Dataframe with columns of phecode description (category) and a mapped a color for node coloring.}

\item{case_size}{Size of nodes drawn for cases (default of 0.1)}

\item{code_size}{Size of nodes drawn for phecodes (default of 0.3)}

\item{no_copies}{Color of cases nodes with no minor allele copy (default grey)}

\item{one_copy}{Color of cases nodes with one minor allele copy (default orangered)}

\item{two_copies}{Color of cases nodes with two minor allele copies (default redish)}
}
\value{
A list of edges and vertices with attributes used by the various network plotting functions included in library
}
\description{
Convert phenotype dataframe to a data list that works with network plotting functions
}
