% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketooltips.R
\name{makeTooltips}
\alias{makeTooltips}
\title{Create a tooltip column for a phewas results dataframe.}
\usage{
makeTooltips(phewas)
}
\arguments{
\item{phewas}{dataframe containing a series of columns that will get concatinated into a single chunk of html with column name followed by the value.}
}
\value{
dataframe with an added tooltip column.
}
\description{
Create a tooltip column for a phewas results dataframe.
}
