% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_manhattan_plot.R
\name{manhattan_plot_UI}
\alias{manhattan_plot_UI}
\title{UI function of manhattan plot module. This module produces a typical manhattan plot with interactive code selection}
\usage{
manhattan_plot_UI(id)
}
\arguments{
\item{id}{String with unique id of module in app}
}
\value{
UI component of interactive manhattan plot
}
\description{
UI function of manhattan plot module. This module produces a typical manhattan plot with interactive code selection
}
\examples{
manhattan_plot_UI('my_mod')
}
