% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergephenomegenome.R
\name{mergePhenomeGenome}
\alias{mergePhenomeGenome}
\title{Merge a tidy phenome dataframe with a tidy genome(single snp) genome
dataframe.}
\usage{
mergePhenomeGenome(phenome_data, genome_data)
}
\arguments{
\item{phenome_data}{Dataframe with columns \code{id} and \code{code}
corresponding to all pairs of case and phenotypes}

\item{genome_data}{Dataframe with columns \code{id} and \code{snp}
corresponding to all pairs of case and snp MA status for a given individual
snp}
}
\value{
A wide dataframe with columns \code{id}, \code{snp}, and all the
unique phecodes.
}
\description{
Merge a tidy phenome dataframe with a tidy genome(single snp) genome
dataframe.
}
\examples{
margePhenomeGenome(tidyPhenome, tidyGenome)
}
