% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergephenomesnp.R
\name{mergePhenomeSnp}
\alias{mergePhenomeSnp}
\title{Merge a wide phenome file and a tidy snp file by IID and desired snp.}
\usage{
mergePhenomeSnp(SNP, phenome_data, snp_data)
}
\arguments{
\item{SNP}{String of desired snp rsid as its encoded in the \code{snp} column of your \code{snp_data} dataframe.}

\item{phenome_data}{Wide phenome dataframe with an \code{IID} column.}

\item{snp_data}{Tidy snp dataframe with \code{IID}, \code{snp}, and \code{copies} columns corresponding to the case ID, snp id, and copies of snp minor allele respectively.}
}
\value{
A wide dataset with columns for case \code{IID}, all the phenotypes, and \code{snp} column corresponding to number of minor allele copies of snp of interest each case has.
}
\description{
Merge a wide phenome file and a tidy snp file by IID and desired snp.
}
\examples{
mergePhenomeSnp('rs9392013', widePhenome, tidySNP)
}
