% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_network2d.R
\name{network2d}
\alias{network2d}
\title{Server function of snp info panel}
\usage{
network2d(input, output, session, network_data, snp_filter,
  viz_type = "free")
}
\arguments{
\item{input, output, session}{Auto-filled by callModule | ignore}

\item{network_data}{Reactive object containing individual network data as generated by \code{meToolkit::makeNetworkData}}

\item{snp_filter}{Reactive object containing boolean containing info on if we've filtered by snp or not.}
}
\description{
Server function of snp info panel
}
\examples{
callModule(info_panel, 'info_panel', snp_name, individual_data, subset_maf)
}
