% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pattern_network.R
\name{pattern_network_UI}
\alias{pattern_network_UI}
\title{UI function of pattern_network module}
\usage{
pattern_network_UI(id, div_class = "pattern_network_plot")
}
\arguments{
\item{id}{String with unique id of module in app}

\item{div_class}{A character string containing a class name for the entire plot to be wrapped in. This can then be used to style with external css. Defaults to 'pattern_network_plot'.}
}
\value{
HTML component of shiny module
}
\description{
UI function of pattern_network module
}
\examples{
pattern_network_UI('my_mod')
}
