% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_phewas_table.R
\name{phewas_table}
\alias{phewas_table}
\title{Server function of phewas results table module. Returns a reactive variable that dispenses interaction events in the common type, payload format of \code{meToolkit} modules.}
\usage{
phewas_table(input, output, session, results_data, selected_codes,
  action_object)
}
\arguments{
\item{input, output, session}{Auto-filled by callModule | ignore}

\item{results_data}{Dataframe containing the results of the phewas study. Needs columns \code{p_val}, \code{id}, \code{category}(along with accompanying \code{color}), \code{tooltip}.}

\item{selected_codes}{A reactive variable containing array of code \code{id}s that are currently selected in the app.}

\item{action_object}{A \code{reactiveVal} that will be updated by the module upon selection}
}
\value{
Shiny module that despenses type-payload actions to the \code{action_object} reactive
}
\description{
Server function of phewas results table module. Returns a reactive variable that dispenses interaction events in the common type, payload format of \code{meToolkit} modules.
}
\examples{
callModule(phewas_table, 'my_mod')
}
