% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_phewas_table.R
\name{phewas_table_UI}
\alias{phewas_table_UI}
\title{UI function of phewas results table module}
\usage{
phewas_table_UI(id)
}
\arguments{
\item{id}{String with unique id of module in app}
}
\value{
HTML component of shiny module
}
\description{
UI function of phewas results table module
}
\examples{
phewas_table_UI('my_mod')
}
