% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_popup.R
\name{pretty_popup}
\alias{pretty_popup}
\title{Reveal hidden element
Targets a given UI element and toggles it from hidden to visable.}
\usage{
pretty_popup(session, title, body)
}
\arguments{
\item{session}{The session variable from your server or module function.}

\item{title}{Character string containing the main title of popup.}

\item{body}{Character string containing the body of popup message.}
}
\value{
A popup over the main UI of your shiny app with desired message.
}
\description{
Reveal hidden element
Targets a given UI element and toggles it from hidden to visable.
}
\examples{
pretty_popup(session, "Oh no", 'You got an error. It was probably because of X,Y, and Z.')
}
