% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile_data.R
\name{reconcile_data}
\alias{reconcile_data}
\title{Combine and reconcile the three main data components needed for ME app and
return in app-friendly format}
\usage{
reconcile_data(phewas_results, id_to_snp, id_to_code)
}
\arguments{
\item{phewas_results}{Dataframe containing the results of the phewas study.
Needs columns \code{p_val}, \code{id}, \code{category}(along with
accompanying \code{color}), \code{tooltip}.}

\item{id_to_snp}{Dataframe containing columns on \code{id}, \code{snp}(#
copies of allele).}

\item{id_to_code}{Dataframe containing column \code{id}, and columns for
each code included.}
}
\value{
List with \code{snp_name}, \code{individual_data}, and \code{phewas_results}
properties for use in ME app.
}
\description{
Helper function used at the begining of a ME app that takes the raw format of
the three data sets required, phewas results, individual snp copies, and
individual phenotype pairs, and combines them into a list object that
combines individual data and checks phewas results along with extracting the
snp name from the column of the genotype file.
}
\examples{
reconcile_data(phewas_table, id_to_snp, phenotype_id_pairs)
}
