% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_individual_data.R
\name{sim_individual_data}
\alias{sim_individual_data}
\title{Generate sample individual level data based upon a PheWAS results set}
\usage{
sim_individual_data(phewas_results, num_patients, snp_prev)
}
\arguments{
\item{phewas_results}{Phewas results. Needs a column \code{code} with the id
of a given phecode, and \code{OR}: the odds ratio of having a code given
having the snp of interest. Optionally, individuals that have the code in
the entire population can be provided with column \code{code_proportion}.}

\item{num_patients}{How many patients should be simulated.}

\item{snp_prev}{The overall population prevalence of the snp you want to
simulate.}
}
\value{
list with following components \describe{ \item{snp_status}{tibble
with patient \code{ID} and snp status (\code{snp})} \item{phenotypes}{wide
tibble with an \code{ID} column and a column for each code provided in the
\code{phewas_resuts} input. Cell values are if the patient had that given
code.} }
}
\description{
This function will take the results of a PheWAS study (most importantly a
code, case and control numbers, and an OR) and simulate as many individual
datapoints as you desire. Should only be used for testing of apps and not any
real simulation studies.
}
\examples{
simulated_phewas <- sim_phewas_results(n_codes = 100, n_categories = 11)
sim_individual_data(simulated_phewas, 10, 0.15)
}
