% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simindividualdata.R
\name{simIndividualData}
\alias{simIndividualData}
\title{Generate sample indivual level data
This function is a helper for prototyping functions. Generates a dataframe of individual-level data for use in other functions.
Note that there is a guarentee that each case will have at least one code, but there is not a guarentee that each code will have a case.}
\usage{
simIndividualData(n_codes, n_cases, prob_snp)
}
\arguments{
\item{n_codes}{How many phecodes needed}

\item{n_cases}{How many cases are needed}

\item{prob_snp}{What's the probability of a copy of a minor allele? A bernouli probability with two trials.}
}
\value{
A dataframe with rows corresponding to each case or patient and columns corresponding to phecode.
In addition, a column \code{IID} for a case id string and \code{snp} corresponding to number of copies of the minor allele are included in output.
}
\description{
Generate sample indivual level data
This function is a helper for prototyping functions. Generates a dataframe of individual-level data for use in other functions.
Note that there is a guarentee that each case will have at least one code, but there is not a guarentee that each code will have a case.
}
