% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_for_main_dashboard.R
\name{subset_to_codes}
\alias{subset_to_codes}
\title{Subset phenotype data to a given list of phecodes}
\usage{
subset_to_codes(data, desired_codes, codes_to_invert = c())
}
\arguments{
\item{data}{A wide phenotype file with columns IID, snp, and all the
normalized phecode names}

\item{desired_codes}{Character vector of the phecodes to be extracted}

\item{codes_to_invert}{Character vector of phecodes that the user has
requested be inverted (e.g. not having the code is considered 'having' it.)}
}
\value{
A wide dataframe with just the columns desired and only rows
corresponding to cases with one or more of the desired codes.
}
\description{
Removes individuals who have none of the requested codes
}
\examples{
subset_to_codes(myPhenotypes, c('001.00', '002.00'), codes_to_invert = c('001.00'))
}
