% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetinteractive_module.R
\name{upset_old}
\alias{upset_old}
\title{Upset Server Function}
\usage{
upset_old(input, output, session, codeData, snpData)
}
\arguments{
\item{input}{Auto-filled by callModule | ignore}

\item{output}{Auto-filled by callModule | ignore}

\item{session}{Auto-filled by callModule | ignore}

\item{codeData}{dataframe containing columns on \code{IID}, \code{snp}(# copies of allele), and columns for each code included.}

\item{snpData}{dataframe containing two columns \code{IID}, \code{snp} for every case in the population. Used in calculating a overall snp abundence.}
}
\description{
Upset Server Function
}
\examples{
callModule(meToolkit::upset, 'upsetPlotV2',
codeData = codeFiltered,
npData = snpData)
}
