% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_popup.R
\name{use_pretty_popup}
\alias{use_pretty_popup}
\title{Use pretty popup function}
\usage{
use_pretty_popup()
}
\value{
Injected javascript function in head of HTML that Shiny uses to send
popup messages to browser.
}
\description{
Function to be included in UI of shiny app that uses \code{\link{pretty_popup}}
function. Injects javascript into HTML.
}
\examples{
shiny::tagList(use_pretty_popup(),h1('Hi there'))
}
\seealso{
\code{\link{pretty_popup}}
}
