% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnaboutselection.R
\name{warnAboutSelection}
\alias{warnAboutSelection}
\title{Warn user about bad selection}
\usage{
warnAboutSelection()
}
\value{
A popup modal that warns the user they made a bad selection (too few codes) and asks them to do it again
}
\description{
Warn user about bad selection
}
\examples{
warnAboutSelection()
}
