% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_info_panel.R
\name{info_panel}
\alias{info_panel}
\title{Server function of snp info panel}
\usage{
info_panel(input, output, session, snp_name, all_individual_data,
  current_individual_data)
}
\arguments{
\item{input, output, session}{Auto-filled by callModule | ignore}

\item{snp_name}{String of the snp name}

\item{all_individual_data}{Individual level data for all patients inc urrent cohort, needs to contain column \code{snp} containing copies of MA.}

\item{current_individual_data}{Reactive individual level data for the currently viewed subset of the cohort, again needs a \code{snp} column.}
}
\value{
Server component of info panel module
}
\description{
Server function of snp info panel
}
\examples{
callModule(info_panel, 'info_panel', snp_name, all_individual_data, current_individual_data)
}
