% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_info_panel.R
\name{info_panel_UI}
\alias{info_panel_UI}
\title{SNP information panel module}
\usage{
info_panel_UI(id)
}
\arguments{
\item{id}{Unique id of module}
}
\value{
UI portion of info panel module
}
\description{
SNP information panel module
}
\examples{

info_panel_UI('myinfo_panel')
}
