% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertcodes.R
\name{invertCodes}
\alias{invertCodes}
\title{Invert currently selected codes Only inverts codes that were not preivously
inverted. If a code is 'inverted' after it's already inverted it simple
reverts back to normal.}
\usage{
invertCodes(codes_to_invert, currently_inverted_codes)
}
\arguments{
\item{codes_to_invert}{Character vector of codes you wish to invert}

\item{currently_inverted_codes}{Character vector of currently inverted codes}
}
\value{
Vector of codes that are currently inverted
}
\description{
Invert currently selected codes Only inverts codes that were not preivously
inverted. If a code is 'inverted' after it's already inverted it simple
reverts back to normal.
}
\examples{
invertCodes(c('001.00', '002.00'), c('001.00', '005.00'))
}
