% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_dashboard.R
\name{main_dashboard}
\alias{main_dashboard}
\title{Main Multimorbidity Explorer Dashboard: Server function}
\usage{
main_dashboard(
  input,
  output,
  session,
  snp_name,
  phewas_results,
  individual_data,
  max_allowed_codes = 40,
  usage_instructions = "default",
  colors = list(light_grey = "#f7f7f7", med_grey = "#d9d9d9", dark_grey = "#bdbdbd",
    light_red = "#fcbba1", dark_red = "#ef3b2c", light_blue = "#4292c6", green =
    "#74c476"),
  debug_mode = FALSE,
  show_back_button_messenger = NULL
)
}
\arguments{
\item{input, output, session}{Auto-filled by callModule | ignore}

\item{snp_name}{Character string containing the RSID of the snp you're
viewing. Used to find annotation information.}

\item{phewas_results}{Dataframe containing the results of the phewas study.
Needs columns \code{p_val}, \code{id}, \code{category}(along with
accompanying \code{color}), \code{tooltip}.}

\item{individual_data}{Dataframe containing columns on \code{IID},
\code{snp}(# copies of allele), and columns for each code included.}

\item{max_allowed_codes}{How many codes can the app show at any given time.
Defaults to 40. (Too many and app may get slow.)}

\item{usage_instructions}{HTML tags corresponding to static content to be
displayed in bottom half of info panel. Any html content works. Defaults to
light description.}

\item{colors}{A list of CSS-valid colors to paint interface in if custom
colors desired. Needs \code{light_grey, med_grey, dark_grey, light_blue,
  light_red, dark_red, light_blue, green}.}

\item{debug_mode}{Boolean controlling if changes in app state should be
recorded in logs. Defaults to off.}
}
\value{
Shiny module of main Multimorbidity Explorer dashboard
}
\description{
Generates a full dashboard page containing various visualizations for
investigating comobidity patterns in individual level data and how they
relate to the results of a phewas analysis.
}
\examples{
callModule(
  main_dashboard, 'main_dashboard',
  my_phewas_results,
  my_individual_data,
  usage_instructions = 'This app is complicated!'
)
}
\seealso{
\code{\link{main_dashboard_UI}}
}
