% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_dashboard.R
\name{main_dashboard_UI}
\alias{main_dashboard_UI}
\title{Main Multimorbidity Explorer Dashboard: UI function}
\usage{
main_dashboard_UI(id, snp_colors = c("#bdbdbd", "#fecc5c", "#a50f15"))
}
\arguments{
\item{id}{String with unique id of module in app}

\item{snp_colors}{Array of valid css color codes for 0, 1, and 2 copies of the
minor allele in network plot.}
}
\value{
HTML component of shiny module
}
\description{
Main Multimorbidity Explorer Dashboard: UI function
}
\examples{
main_dashboard_UI('my_app', snp_colors = c('#bdbdbd','#fecc5c', '#a50f15'))
}
\seealso{
\code{\link{main_dashboard}}
}
