% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_manhattan_and_table.R
\name{manhattan_plot_and_table_UI}
\alias{manhattan_plot_and_table_UI}
\title{Interactive manhattan plot and table}
\usage{
manhattan_plot_and_table_UI(id)
}
\arguments{
\item{id}{String with unique id of module in app}
}
\value{
UI component of interactive manhattan plot
}
\description{
For selecting codes for further visualization in app.
}
\examples{
manhattan_plot_and_table_UI('my_mod')
}
\seealso{
\code{\link{manhattan_plot_and_table}}
}
