% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_network2d.R
\name{network2d_UI}
\alias{network2d_UI}
\title{Interactive individual-level data network diagram of patients to codes}
\usage{
network2d_UI(id, height = "500px")
}
\arguments{
\item{id}{Unique id of module}

\item{height}{How tall we want this module to be in css units (defaults to '500px')}
}
\value{
HTML tag containing interactive network
}
\description{
Interactive individual-level data network diagram of patients to codes
}
\examples{

network2d_UI('mycomorbidityNetwork2d', '100\%')
}
