% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_network_plot.R
\name{network_plot_UI}
\alias{network_plot_UI}
\title{Subject-level network diagram: UI}
\usage{
network_plot_UI(id, snp_colors)
}
\arguments{
\item{id}{Unique id of module}

\item{snp_colors}{A three element array of colors corresponding to the color
of patient nodes and their snp status in order of 0,1,2 copies of minor
allele.}
}
\value{
HTML tag containing interactive network
}
\description{
Subject-level network diagram: UI
}
\examples{

network_plot_UI('mycomorbiditynetwork_plot', snp_colors = c('#bdbdbd','#fecc5c', '#a50f15'))
}
\seealso{
\code{\link{network_plot}}
}
