% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_me.R
\name{run_me}
\alias{run_me}
\title{Run ME app locally}
\usage{
run_me()
}
\value{
Shiny app process
}
\description{
Function to spawn a shiny app instance for Multimorbidity Explorer app with
data loading screen.
}
\examples{
\dontrun{
meToolkit::run_me()
}
}
