% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_me.R
\name{run_me}
\alias{run_me}
\title{Run a ME app with data loading screen}
\usage{
run_me(
  preloaded_path = NULL,
  auto_run = FALSE,
  snp_colors = c("#bdbdbd", "#fecc5c", "#a50f15")
)
}
\arguments{
\item{preloaded_path}{File path relative to app that preloaded data is
stored. Defaults to \code{NULL}.}

\item{auto_run}{Do you want the app to run immediately or do you want the ui
and server components of the app to run later? Defaults to \code{TRUE}.}

\item{snp_colors}{Array of valid css color codes for 0, 1, and 2 copies of the
minor allele in network plot.}
}
\value{
Shiny app process
}
\description{
Function to spawn a shiny app instance for Multimorbidity Explorer app with
data loading screen. Optionally data can be preloaded.
}
\examples{
\dontrun{
meToolkit::run_me()
}
}
