% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upset.R
\name{upset}
\alias{upset}
\title{Server function of upset module}
\usage{
upset(input, output, session, individual_data, all_patient_snps,
  action_object = NULL)
}
\arguments{
\item{input, output, session}{Auto-filled by callModule | ignore}

\item{individual_data}{Reactive object with dataframe containing columns on \code{IID}, \code{snp}(# copies of allele), and columns for each code included.}

\item{all_patient_snps}{dataframe containing two columns \code{IID}, \code{snp} for every case in the population. Used in calculating a overall snp abundence.}

\item{action_object}{A \code{reactiveVal} that will be updated by the module upon isolation, deletion, or snp_filtering.}
}
\value{
Shiny module
}
\description{
Server function of upset module
}
\examples{
callModule(upset, 'my_mod')
}
