% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetinteractive_module.R
\name{upset_UI}
\alias{upset_UI}
\title{Upset UI function}
\usage{
upset_UI(id, app_ns = I)
}
\arguments{
\item{id}{ID of shiny module. Needs to match the server-side call.}

\item{app_ns}{A namespace function if this module is included within another module.}
}
\description{
Upset UI function
}
\examples{
upset_ui('upsetPlot', parentNS)
}
