% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upset.R
\name{upset_UI}
\alias{upset_UI}
\title{Upset plot of multimorbidity patterns: UI}
\usage{
upset_UI(id)
}
\arguments{
\item{id}{String with unique id of module in app}
}
\value{
HTML component of shiny module
}
\description{
Upset plot of multimorbidity patterns: UI
}
\examples{
upset_UI('my_mod')
}
\seealso{
\code{\link{upset}}
}
