% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upset.R
\name{upset_UI}
\alias{upset_UI}
\title{UI function of upset module}
\usage{
upset_UI(id)
}
\arguments{
\item{id}{String with unique id of module in app}
}
\value{
HTML component of shiny module
}
\description{
UI function of upset module
}
\examples{
upset_UI('my_mod')
}
